+++ lib/duckdb_dart.dart
@@ -2,7 +2,6 @@
 ///
 library duckdb;
 
-export 'src/duckdb_base.dart';
 export 'src/ffi/duckdb.g.dart';
 export 'src/ffi/load_library.dart';
 export 'src/implementation/duckdb.dart';