+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 
 ## 2024-05-21 (0.1.1) 
-- Support more DuckDb numeric types
+- Support more DuckDb numeric types 
+- Add a database `Configuration` argument to `Connection`
 
 
 ## 2024-05-20 (0.1.0)